Simple 2D Graphics Engine
The task was to write a 2D graphics engine for drawing composite objects
out of basic primitives. The program reads in a file containing a
description of a scene in terms of n sided polygons and circles along
with 2 basic types of manipulation (a 2D translation and a 2D scale).
The program then turns this description into a 2D line drawing.